Update README.md
# Sweater Weather

This is the final solo project for Turing's Back End Engineering Course, Module 3. This project consumes multiple endpoints and presents data to a wireframed front end. The data consumed includes an endpoint for weather, directions, cities, and even allows for the creation of road trips. 

<div align="center">
  <kbd>
    <img src="https://i.imgur.com/zdY3ZC9.png" />
  </kbd>
</div>

## Description
This project consumes weather data for a city, with a current, hourly, and daily set of weather attributes. It allows a front end to plot weather and city information, and